+++ packages/driver/cypress/integration/cypress/browser_spec.js
@@ -1,4 +1,4 @@
-const browserProps = require('@packages/driver/src/cypress/browser')
+import browserProps from '@packages/driver/src/cypress/browser'
 
 describe('src/cypress/browser', () => {
   beforeEach(function () {