+++ packages/driver/src/cy/net-stubbing/events/before-request.ts
@@ -1,20 +1,21 @@
 import _ from 'lodash'
 
 import {
+  SERIALIZABLE_REQ_PROPS,
   Interception,
   CyHttpMessages,
-  SERIALIZABLE_REQ_PROPS,
   Subscription,
 } from '../types'
+
 import { parseJsonBody, stringifyJsonBody } from './utils'
 import {
   validateStaticResponse,
   parseStaticResponseShorthand,
 } from '../static-response-utils'
-import * as $errUtils from '../../../cypress/error_utils'
-import { HandlerFn, HandlerResult } from '.'
+import $errUtils from '../../../cypress/error_utils'
+import type { HandlerFn, HandlerResult } from '.'
 import Bluebird from 'bluebird'
-import { NetEvent } from '@packages/net-stubbing/lib/types'
+import type { NetEvent } from '@packages/net-stubbing/lib/types'
 import Debug from 'debug'
 
 const debug = Debug('cypress:driver:net-stubbing:events:before-request')