+++ packages/server/lib/server-base.ts
@@ -7,28 +7,30 @@ import express, { Express } from 'express'
 import http from 'http'
 import httpProxy from 'http-proxy'
 import _ from 'lodash'
-import { AddressInfo } from 'net'
+import type { AddressInfo } from 'net'
 import url from 'url'
 import la from 'lazy-ass'
-import httpsProxy from '@packages/https-proxy'
+import type httpsProxy from '@packages/https-proxy'
 import { netStubbingState, NetStubbingState } from '@packages/net-stubbing'
 import { agent, clientCertificates, cors, httpUtils, uri } from '@packages/network'
 import { NetworkProxy, BrowserPreRequest } from '@packages/proxy'
-import { SocketCt } from '@packages/server-ct'
+import type { SocketCt } from './socket-ct'
 import errors from './errors'
 import logger from './logger'
 import Request from './request'
-import { SocketE2E } from './socket-e2e'
+import type { SocketE2E } from './socket-e2e'
 import templateEngine from './template_engine'
 import { ensureProp } from './util/class-helpers'
 import origin from './util/origin'
 import { allowDestroy, DestroyableHttpServer } from './util/server_destroy'
 import { SocketAllowed } from './util/socket_allowed'
 import { createInitialWorkers } from '@packages/rewriter'
-import { SpecsStore } from './specs-store'
-import { InitializeRoutes } from '../../server-ct/src/routes-ct'
-import { Cfg } from './project-base'
-import { Browser } from '@packages/server/lib/browsers/types'
+import type { SpecsStore } from './specs-store'
+import type { Cfg } from './project-base'
+import type { Browser } from '@packages/server/lib/browsers/types'
+import { InitializeRoutes, createCommonRoutes } from './routes'
+import { createRoutesE2E } from './routes-e2e'
+import { createRoutesCT } from './routes-ct'
 
 const ALLOWED_PROXY_BYPASS_URLS = [
   '/',
@@ -100,7 +102,6 @@ export interface OpenServerOptions {
   getCurrentBrowser: () => Browser
   getSpec: () => Cypress.Cypress['spec'] | null
   shouldCorrelatePreRequests: () => boolean
-  createRoutes: (args: InitializeRoutes) => any
 }
 
 export abstract class ServerBase<TSocket extends SocketE2E | SocketCt> {
@@ -174,7 +175,6 @@ export abstract class ServerBase<TSocket extends SocketE2E | SocketCt> {
     onWarning,
     shouldCorrelatePreRequests,
     specsStore,
-    createRoutes,
     testingType,
     SocketCtor,
   }: OpenServerOptions) {
@@ -211,8 +211,7 @@ export abstract class ServerBase<TSocket extends SocketE2E | SocketCt> {
 
       this.createHosts(config.hosts)
 
-      createRoutes({
-        app,
+      const routeOptions: InitializeRoutes = {
         config,
         specsStore,
         getRemoteState,
@@ -221,7 +220,15 @@ export abstract class ServerBase<TSocket extends SocketE2E | SocketCt> {
         onError,
         getSpec,
         getCurrentBrowser,
-      })
+        testingType,
+      }
+
+      const runnerSpecificRouter = testingType === 'e2e'
+        ? createRoutesE2E(routeOptions)
+        : createRoutesCT(routeOptions)
+
+      app.use(runnerSpecificRouter)
+      app.use(createCommonRoutes(routeOptions))
 
       return this.createServer(app, config, onWarning)
     })
@@ -394,7 +401,7 @@ export abstract class ServerBase<TSocket extends SocketE2E | SocketCt> {
     })
   }
 
-  _getRemoteState () {
+  _getRemoteState (): Cypress.RemoteState {
     // {
     //   origin: "http://localhost:2020"
     //   fileServer:
@@ -422,7 +429,7 @@ export abstract class ServerBase<TSocket extends SocketE2E | SocketCt> {
       visiting: this._remoteVisitingUrl,
       domainName: this._remoteDomainName,
       fileServer: this._remoteFileServer,
-    })
+    }) as Cypress.RemoteState
 
     debug('Getting remote state: %o', props)